requirements.txt
appdirs==1.4.3
click==6.7
coverage==4.3.4
dj-database-url==0.4.2
Django==1.10
django-filter==1.0.2
djangorestframework==3.6.2
gunicorn==19.7.1
Markdown==2.6.8
packaging==16.8
psycopg2==2.7.1
py==1.4.33
pyparsing==2.2.0
python-dotenv==0.6.4
six==1.10.0
whitenoise==3.3.0
